pyproject.toml
[tool.poetry]
name = "hades"
version = "0.1.0"
description = "Heterogeneous Adaptive Dimensional Embedding System - Production infrastructure for processing and serving embedded knowledge"
authors = ["HADES Team"]
readme = "README.md"
license = "Apache-2.0"
# Installable packages from this repo
packages = [
    { include = "core" },
    { include = "tools" },
]

[tool.poetry.dependencies]
python = "^3.11"
python-arango = "^7.6.0"
numpy = "^1.24.0"
torch = "^2.0.0"
transformers = "^4.35.0"
tqdm = "^4.65.0"
pypdfium2 = "^4.18.0"
pillow = "^10.0.0"
python-magic = "^0.4.27"
docling = "^1.0.0"
pandas = "^2.0.0"
lmdb = "^1.4.0"
sentence-transformers = "^2.2.0"
matplotlib = "^3.7.0"
httpx = "^0.24.0"
pyyaml = "^6.0"
# mcp = { version = "^0.1.0", optional = true }  # Uncomment when MCP is available
aiofiles = "^23.0.0"
einops = "^0.8.1"
orjson = "^3.11.1"
gputil = "^1.4.0"
tenacity = "^9.1.2"
ray = "^2.48.0"
peft = "^0.17.0"
structlog = "^25.4.0"
pymupdf = "^1.26.4"
psycopg = {version = "^3.1.18", extras = ["binary"]}
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
kagglehub = "^0.3.13"
# Graph Neural Network dependencies
torch-geometric = "^2.6.1"
# NOTE: PyG sparse dependencies must be installed separately from PyG wheels repo:
# pip install pyg-lib torch-scatter torch-sparse torch-cluster torch-spline-conv -f https://data.pyg.org/whl/torch-2.7.0+cu126.html
# Version pins removed for sparse libs as they're managed by PyG wheels
scikit-learn = "^1.3.0"
wandb = "^0.16.0"
pynvml = "^11.5.0"
networkx = "^3.2.0"
faiss-cpu = {version = "^1.12.0", python = ">=3.11,<3.15"}
python-json-logger = "^3.3.0"
python-dotenv = "^1.1.1"
watchdog = "^6.0.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.5.0"
black = "^23.0.0"
ruff = "^0.1.0"
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"
types-tqdm = "*"
types-pillow = "*"

# [tool.poetry.extras]
# mcp = ["mcp"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_calls = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

# Per-module options
[[tool.mypy.overrides]]
module = "transformers.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "torch.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "arango.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tqdm.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "numpy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pypdfium2.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "torch_geometric.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sklearn.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "wandb.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pynvml.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "networkx.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "PIL.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "yaml"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "psycopg.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sqlalchemy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "alembic.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sickle.*"
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 120
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 120
target-version = ["py311"]
